Move profile models to models.py
