+++ backend/btrixcloud/profiles.py
@@ -1,86 +1,32 @@
 """ Profile Management """
 
-from typing import Optional, List
+from typing import Optional
 from datetime import datetime
 import uuid
 import os
 
 from urllib.parse import urlencode
 
+from pydantic import UUID4
 from fastapi import APIRouter, Depends, Request, HTTPException
-from pydantic import BaseModel, UUID4, HttpUrl
 import aiohttp
 
 from .pagination import DEFAULT_PAGE_SIZE, paginated_format
 from .users import User
 
-from .db import BaseMongoModel
-from .models import CrawlConfigIdNameOut, Organization
+from .models import (
+    Profile,
+    ProfileWithCrawlConfigs,
+    ProfileFile,
+    UrlIn,
+    ProfileLaunchBrowserIn,
+    BrowserId,
+    ProfileCreateUpdate,
+    Organization,
+)
 
 
 BROWSER_EXPIRE = 300
-
-
-# ============================================================================
-class ProfileFile(BaseModel):
-    """file from a crawl"""
-
-    filename: str
-    hash: str
-    size: int
-
-
-# ============================================================================
-class Profile(BaseMongoModel):
-    """Browser profile"""
-
-    name: str
-    description: Optional[str] = ""
-
-    userid: UUID4
-    oid: UUID4
-
-    origins: List[str]
-    resource: Optional[ProfileFile]
-
-    created: Optional[datetime]
-
-
-# ============================================================================
-class ProfileWithCrawlConfigs(Profile):
-    """Profile with list of crawlconfigs using this profile"""
-
-    crawlconfigs: List[CrawlConfigIdNameOut] = []
-
-
-# ============================================================================
-class UrlIn(BaseModel):
-    """Request to set url"""
-
-    url: HttpUrl
-
-
-# ============================================================================
-class ProfileLaunchBrowserIn(UrlIn):
-    """Request to launch new browser for creating profile"""
-
-    profileId: Optional[UUID4]
-
-
-# ============================================================================
-class BrowserId(BaseModel):
-    """Profile id on newly created profile"""
-
-    browserid: str
-
-
-# ============================================================================
-class ProfileCreateUpdate(BaseModel):
-    """Profile metadata for committing current browser to profile"""
-
-    browserid: Optional[str]
-    name: str
-    description: Optional[str] = ""
 
 
 # ============================================================================
